requirements.txt
openai
pypdf
faiss-cpu
langchain
langchain-community
langchain_openai
fastapi
uvicorn
python-dotenv
